tests/wpt/web-platform-tests/largest-contentful-paint/loadTime-after-appendChild.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: delayed appended image.</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<script>
  async_test(function (t) {
    assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
    let beforeLoad;
    const observer = new PerformanceObserver(
      t.step_func_done(entryList => {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const url = window.location.origin + '/images/black-rectangle.png';
        // blue.png is 100 by 50.
        const size = 100 * 50;
        checkImage(entry, url, 'image_id', size, beforeLoad);
      })
    );
    observer.observe({type: 'largest-contentful-paint', buffered: true});
    const img = document.createElement('img');
    img.src = '/images/black-rectangle.png';
    img.id = 'image_id';
    t.step_timeout(() => {
      beforeLoad = performance.now();
      document.getElementById('image_div').appendChild(img);
    }, 200)
  }, 'Image loadTime occurs after appendChild is called.');
</script>
<div id='image_div'></div>
</body>
